+++ nimare/meta/__init__.py
@@ -4,5 +4,6 @@ Top-level namespace for meta-analyses.
 
 from . import cbma
 from . import ibma
+from . import esma
 
-__all__ = ['cbma', 'ibma']
+__all__ = ['cbma', 'ibma', 'esma']
