+++ nimare/meta/cbma/ale.py
@@ -157,7 +157,7 @@ class ALE(CBMAEstimator):
             ma_maps = k_est.transform(self.ids, **self.kernel_arguments)
             images = self._run_ale(ma_maps, prefix='', n_cores=n_cores)
 
-        self.results = MetaResult(self, mask=self.mask, **images)
+        self.results = MetaResult(self, self.mask, maps=images)
 
     def subtraction_analysis(self, ids, ids2, image1, image2, ma_maps=None):
         """
@@ -654,7 +654,7 @@ class SCALE(CBMAEstimator):
                   'p': p_values,
                   'z': z_values,
                   'vthresh': vthresh_z_values}
-        self.results = MetaResult(self, mask=self.mask, **images)
+        self.results = MetaResult(self, self.mask, maps=images)
 
     def _compute_ale(self, df=None, ma_maps=None):
         """
