+++ nimare/tests/test_esma.py
@@ -1,5 +1,5 @@
 """
-Test nimare.meta.ibma (image-based meta-analytic algorithms).
+Test nimare.meta.esma (effect-size meta-analytic algorithms).
 """
 import json
 import os.path as op
@@ -10,7 +10,7 @@ import nibabel as nib
 from nilearn.masking import apply_mask
 
 import nimare
-from nimare.meta import ibma
+from nimare.meta import esma
 from nimare.stats import t_to_z
 from nimare.tests.utils import get_test_data_path, download_nidm_pain
 
@@ -136,60 +136,57 @@ def test_z_perm():
     """
     Smoke test for z permutation.
     """
-    result = ibma.stouffers(pytest.z_data, pytest.mask_img,
-                            inference='rfx', null='empirical', n_iters=10,
-                            corr='FDR')
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.stouffers(pytest.z_data, inference='rfx', null='empirical',
+                            n_iters=10, corr='FDR')
+    assert isinstance(result, dict)
 
 
 def test_stouffers_ffx():
     """
     Smoke test for Stouffer's FFX.
     """
-    result = ibma.stouffers(pytest.z_data, pytest.mask_img,
-                            inference='ffx', null='theoretical', n_iters=None)
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.stouffers(pytest.z_data, inference='ffx', null='theoretical',
+                            n_iters=None)
+    assert isinstance(result, dict)
 
 
 def test_stouffers_rfx():
     """
     Smoke test for Stouffer's RFX.
     """
-    result = ibma.weighted_stouffers(pytest.z_data, pytest.sample_sizes_z,
-                                     pytest.mask_img)
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.weighted_stouffers(pytest.z_data, pytest.sample_sizes_z)
+    assert isinstance(result, dict)
 
 
 def test_weighted_stouffers():
     """
     Smoke test for Weighted Stouffer's.
     """
-    result = ibma.stouffers(pytest.z_data, pytest.mask_img,
-                            inference='rfx', null='theoretical', n_iters=None)
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.stouffers(pytest.z_data, inference='rfx', null='theoretical',
+                            n_iters=None)
+    assert isinstance(result, dict)
 
 
 def test_fishers():
     """
     Smoke test for Fisher's.
     """
-    result = ibma.fishers(pytest.z_data, pytest.mask_img)
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.fishers(pytest.z_data)
+    assert isinstance(result, dict)
 
 
 def test_con_perm():
     """
     Smoke test for contrast permutation.
     """
-    result = ibma.rfx_glm(pytest.con_data, pytest.mask_img, null='empirical',
-                          n_iters=10, corr='FDR')
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.rfx_glm(pytest.con_data, null='empirical', n_iters=10,
+                          corr='FDR')
+    assert isinstance(result, dict)
 
 
 def test_rfx_glm():
     """
     Smoke test for RFX GLM.
     """
-    result = ibma.rfx_glm(pytest.con_data, pytest.mask_img, null='theoretical',
-                          n_iters=None)
-    assert isinstance(result, nimare.base.meta.MetaResult)
+    result = esma.rfx_glm(pytest.con_data, null='theoretical', n_iters=None)
+    assert isinstance(result, dict)
